chore: clean up code and namings
